# documentation/openapi.yaml
openapi: '3.0.3'
info:
  title: OpenEvents
  description: An Express API for OpenEvents
  contact:
    email: danimr1999@gmail.com
    name: Daniel Muelle Ramos
  license:
    name: MIT
    url: https://www.mit.edu/~amini/LICENSE.md
  version: 1.0.0
servers:
- url: http://localhost:3000/
  description: Development
paths:
...
  /friendships/requests:
    get:
      tags:
        - Friendships
      summary: Gets all the users that have sent a friendship request to the authenticated user and its status is still pending.
      security: 
        - BearerAuth: []
      responses:
        "200":
          description: List of users that have sent a frienship request to the authenticated user and its status is still pending.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseListUserWithId'
  /friendships:
    get:
      tags:
        - Friendships
      summary: Gets all friends of the authenticated user.
      security: 
        - BearerAuth: []
      responses:
        "200":
          description: List of friends of the authenticated user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseListUserWithId'
  /friendships/{user_id}:
    post:
      tags:
        - Friendships
      summary: Creates a friend request to an external user from an authenticated user.
      description: Creates a friend request to an external user. In case the external user has already sent a 
        friend request to the authenticated user, it is accepted.
      security: 
        - BearerAuth: []
      parameters:
          - in: path
            name: user_id
            schema:
              type: integer
            required: true
            description: ID of the external user
      responses:
        "200":
          description: Message notification informing whether users were already friends, friend request was accepted or
            was already sent.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  stacktrace:
                    type: object
        "201":
          description: Message notification informing that a friend request has been sent.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  stacktrace:
                    type: object
        "400":
          description: Authenticated user has tried to send a friend request to itself.
        "404":
          description: External user does not exist or was not found.
    put:
      tags:
        - Friendships
      summary: Accepts a friend request from an external user to an authenticated user.
      security: 
        - BearerAuth: []
      parameters:
          - in: path
            name: user_id
            schema:
              type: integer
            required: true
            description: ID of the external user
      responses:
        "200":
          description: Message notification informing whether users were already friends, friend request was accepted or
            friend request must be accepted by the external user.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  stacktrace:
                    type: object
        "400":
          description: Authenticated user has tried to accept a friend request from itself.
        "404":
          description: External user does not exist or was not found.
    delete:
      tags:
        - Friendships
      summary: Deletes a friend request between the authenticated user and the external user.
      description: Used to delete a friend request regardless of its status or a friendship (when a friend request is accepted).
        In case exists, it also deletes all the messages exchanged between both users.
      security: 
        - BearerAuth: []
      parameters:
          - in: path
            name: user_id
            schema:
              type: integer
            required: true
            description: ID of the external user
      responses:
        "200":
          description: Message notification informing whether friendship or friend request was deleted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  stacktrace:
                    type: object
        "400":
          description: Authenticated user has tried to delete a friend request with itself.
        "404":
          description: External user does not exist or was not found.
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ResponseListUserWithId:
      type: array
      items:
        type: object
        properties:
            id:
              type: number
              example: 1
            name:
              type: string
              example: "Jane"
            last_name:
              type: string
              example: "Doe"
            email:
              type: string
              example: "janedoe@openevents.com"
            image_url:
              type: string
              example: "https://i.imgur.com/ghy8Xx1.png"
...